Rename property.rs to media_type.rs
